+++ src/guides/legend.jl
@@ -380,9 +380,8 @@ end
 datavalues_plotvalues_datalabels(aes, scale::CategoricalScale) = datavalues(scale), plotvalues(scale), datalabels(scale)
 function datavalues_plotvalues_datalabels(aes::Type{AesMarkerSize}, scale::ContinuousScale)
     props = scale.props.aesprops::AesMarkerSizeContinuousProps
-    tickfinder = temporal_tickfinder(scale, props.ticks)
-    scale = strip_temporal(scale)
-    _, s_extrema = strip_units(scale, collect(nonsingular_limits(scale.extrema)))
+    scale, tickfinder = strip_scale_derive_ticks(scale, props.ticks)
+    s_extrema = nonsingular_limits(scale.extrema)
     tickvalues, ticklabels = Makie.get_ticks(tickfinder, identity, props.tickformat, s_extrema...)
     t_extrema = extrema(tickvalues)
     if t_extrema[1] < s_extrema[1] || t_extrema[2] > s_extrema[2]
@@ -393,9 +392,8 @@ function datavalues_plotvalues_datalabels(aes::Type{AesMarkerSize}, scale::Conti
 end
 function datavalues_plotvalues_datalabels(aes::Type{AesLineWidth}, scale::ContinuousScale)
     props = scale.props.aesprops::AesLineWidthContinuousProps
-    tickfinder = temporal_tickfinder(scale, props.ticks)
-    scale = strip_temporal(scale)
-    _, s_extrema = strip_units(scale, collect(nonsingular_limits(scale.extrema)))
+    scale, tickfinder = strip_scale_derive_ticks(scale, props.ticks)
+    s_extrema = nonsingular_limits(scale.extrema)
     tickvalues, ticklabels = Makie.get_ticks(tickfinder, identity, props.tickformat, s_extrema...)
     t_extrema = extrema(tickvalues)
     if t_extrema[1] < s_extrema[1] || t_extrema[2] > s_extrema[2]